+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#26](https://github.com/LaxarJS/laxar-dox/issues/26): switched to swig-templates
 - [#27](https://github.com/LaxarJS/laxar-dox/issues/27): added support for ES2015 class syntax