+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.1.0
+
 - [#25](https://github.com/LaxarJS/laxar-dox/issues/25): added support for deprecated tag
     + NEW FEATURE: see ticket for details
 - [#26](https://github.com/LaxarJS/laxar-dox/issues/26): switched to swig-templates